test(file_archiver): split reproducibility and golden hashes integration tests into multiple files
