+++ internal/mithril-file-archiver/tests/extensions/mod.rs
@@ -1,0 +1,6 @@
+// Avoid clippy warnings generated by tests that doesn't use every tests_extensions (since each test
+// is a different compilation target).
+#![allow(dead_code)]
+
+pub mod helpers;
+pub mod test_data;
